:sparkles: feat : 디렉토리 구조 변경, 에러/로딩/결과 없을 시 적절한 UI 생성 #162
