+++ frontend/src/components/search/Search.tsx
@@ -1,9 +1,10 @@
 import { useEffect, useState } from 'react';
 import SearchModal from './SearchModal';
+import '../../style/Search.css';
 
 function Search() {
   const [showSearch, setShowSearch] = useState<boolean>(false);
-  
+
   function handleKeydown(e: KeyboardEvent) {
     if (e.key === 'k' && e.metaKey) {
       setShowSearch(true);
@@ -16,12 +17,13 @@ function Search() {
       document.removeEventListener('keydown', handleKeydown);
     };
   }, []);
-  
-
 
   return (
     <div className="search">
-      <button className="search-btn" onClick={() => setShowSearch(!showSearch)}>
+      <button
+        className="searchButton"
+        onClick={() => setShowSearch(!showSearch)}
+      >
         안녕 여기는 검색 버튼 이란다
       </button>
       {showSearch && <SearchModal setShowSearch={setShowSearch} />}
